+++ acorn/scipy/__init__.py
@@ -4,9 +4,6 @@
 # explicitly asked for before they are loaded into the package namespace. So we
 # do that here.
 
-#Numpy is almost completely included in scipy; if we want the arrays etc. to be
-#handled correctly, we ought to decorate that here as well.
-import acorn.numpy
 from acorn.logging.decoration import set_decorating, decorating
 
 #Before we do any imports, we need to set that we are decorating so that
@@ -19,6 +16,12 @@ from scipy import optimize, spatial, stats, signal, odr, io, constants
 from scipy import cluster 
 from acorn.logging.decoration import decorate
 decorate(asp)
+
+#Because scipy uses a lot of numpy, we run into problems where top-level scipy
+#members point to the *undecorated* numpy objects. Here we just ensure that all
+#of scipy is decorated.
+from acorn.logging.decoration import postfix
+postfix(asp)
 
 import sys
 sys.modules[__name__] = asp